requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
DateTime==4.3
Django==3.0.5
heroku==0.1.4
idna==2.9
joblib==0.14.1
naive-bayes==0.1.1
numpy==1.18.2
pandas==1.0.3
python-dateutil==1.5
python-firebase==1.2
pytz==2019.3
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
urllib3==1.25.8
zope.interface==5.1.0
